multicast-menu/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.7
certifi==2024.7.4
cffi==1.17.0rc1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==35.0.0
Deprecated==1.2.14
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.0.0
gunicorn==20.1.0
idna==3.7
importlib_metadata==7.1.0
ipwhois==1.2.0
kombu==5.4.0rc2
Pillow==9.0.1
prompt_toolkit==3.0.47
psutil==5.8.0
psycopg2-binary==2.9.9
pycparser==2.22
pytz==2024.1
redis==4.0.2
requests==2.26.0
scapy==2.4.5
sqlparse==0.5.1
typing_extensions==4.12.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
wrapt==1.16.0
zipp==3.19.2
